+++ PvpAutoLb/Core/Filters/StatusFilter.cs
@@ -1,0 +1,16 @@
+using Dalamud.Game.ClientState.Objects.Types;
+
+namespace PvpAutoLb.Core;
+
+internal static class StatusFilter
+{
+    public static bool IsGuarded(IBattleChara target)
+    {
+        var list = target.StatusList;
+        for (var i = 0; i < list.Length; i++)
+        {
+            if (list[i] is { StatusId: PvpAutoLbConstants.StatusIds.Guard }) return true;
+        }
+        return false;
+    }
+}